People Also Ask Scraper
This Python script uses Selenium to extract "People also ask" questions from Google search results for each keyword specified in a text file. If a CAPTCHA is encountered, the script waits for you to manually solve it in the browser. The number of questions to extract per query is configurable via the settings file.